+++ packages/components/src/Chips/InternalChip.css.d.ts
@@ -1,3 +1,4 @@
+export const wrapper: string;
 export const chip: string;
 export const hasPrefix: string;
 export const hasSuffix: string;
