--- packages/components/src/Chips/InternalChipChoice.css
@@ -1,5 +1,0 @@
-.wrapper {
-  display: flex;
-  gap: var(--space-small);
-  flex-wrap: wrap;
-}
