--- packages/components/src/Chips/InternalChipChoice.css.d.ts
@@ -1,1 +1,0 @@
-export const wrapper: string;
